Add marathon_task_states metrics.
Existing marathon metrics are population orientated, rather than individual
orientated- if I want to track what's going on with one particular task
(or map all tasks back to a marathon app), there is no linkage.

Thus this metric: it's duplicative for the others (to preserve compatibility),
but this links the state of a task, the app it's from, the slaveId it's on,
the config version, and the state.  This gives full drill down visibility
into the progressions of a task.

This can be used for the following for example:
* linking cadvisor level metrics back to the marathon task itself, linking
  through mesos metrics for joining.
* isolating performance variants of a given task based on the node they
  were running on (comparing old and new gen hardware for example).
* correlating task launch failures to specific nodes on an app by app
  basis.  There are other metrics to identify bad nodes- but this is another
  way to attack it.