Add -app.label.default to allow supplying default application labels.
This is something of an edge case usage; if you are exporting thresholds
as a label, but want to enforce a default threshold- IE, a default label-
there is no way to do this without relabeling every app in a given marathon.

Thus `-app.label.default`; this can be used to inject a default label into
the collected marathon application labels.  If the app defines the label
itself, then that value is used.  If not, the default is used.

This flag- in conjunction with the other -app.label.* flags- allows synthesizing
a metric for all applications in marathon, while allowing override on a per
app basis.  That in turn makes promql interactions (including alerting if it's
a threshold) saner to write.